+++ src/api/estimate-coin-sell.js
@@ -22,7 +22,7 @@ export default function EstimateCoinSell(apiInstance) {
      * @param {string|number} params.valueToSell
      * @param {string|number} params.coinToBuy - ID or symbol of the coin to buy
      * @param {ESTIMATE_SWAP_TYPE} [params.swapFrom] - estimate pool swap
-     * @param {Array<string|number>} [params.route] - intermediate coins IDs for pool swaps
+     * @param {Array<string|number>} [params.route] - IDs of intermediate coins for pool swaps
      * @param {string|number} [params.gasCoin]
      * @param {string|number} [params.coinCommission] - gasCoin alias
      * @param {import('axios').AxiosRequestConfig} [axiosOptions]