+++ src/index.js
@@ -7,13 +7,12 @@ export {default as PostTx, EnsureNonce} from './api/post-tx.js';
 export {default as PostSignedTx} from './api/post-signed-tx.js';
 export {default as GetNonce} from './api/get-nonce.js';
 export {default as GetCoinInfo} from './api/get-coin-info.js';
-export {default as GetCoinInfoById} from './api/get-coin-info-by-id.js';
 export {default as GetMinGasPrice} from './api/get-min-gas-price.js';
 export {default as EstimateCoinSell} from './api/estimate-coin-sell.js';
 export {default as EstimateCoinSellAll} from './api/estimate-coin-sell-all.js';
 export {default as EstimateCoinBuy} from './api/estimate-coin-buy.js';
 export {default as EstimateTxCommission} from './api/estimate-tx-commission.js';
-export {ReplaceCoinSymbol, ReplaceCoinSymbolByPath} from './api/replace-coin.js';
+export {ReplaceCoinSymbol, ReplaceCoinSymbolByPath, GetCoinId, ReplaceCoinId, ReplaceCoinIdByPath, GetCoinSymbol} from './api/replace-coin.js';
 export {default as issueCheck, decodeCheck, getGasCoinFromCheck} from './check.js';
 export {prepareLink, decodeLink} from './link.js';